requirements.txt
asgiref==3.5.0
azure-core==1.23.1
azure-storage-blob==12.11.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.3
django-allauth==0.49.0
django-request==1.6.1
django-storages==1.12.3
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
gunicorn==20.1.0
idna==3.3
isodate==0.6.1
msrest==0.6.21
oauthlib==3.2.0
Pillow==9.1.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
